refactoring: 기획 변경으로인한 채팅방 생성, 삭제 주석처리
