+++ src/app/(auth)/(navigationsBarLayout)/@chatDialog/page.tsx
@@ -1,7 +1,7 @@
 import { Dialog, DialogContent } from '@/components/ui/dialog';
 import { DialogTitle } from '@radix-ui/react-dialog';
 import ChatRoomList from '@/features/chat/ui/ChatRoomList';
-import ChatRoom from '@/features/chat/ui/ChatRoom';
+import ChatRoom from '@/features/chat/ui/chatRoom';
 
 interface ChatDialogProps {
   searchParams: Promise<{ 'room-id': string; 'chat-open': boolean }>;
@@ -13,7 +13,7 @@ export default async function ChatDialog({ searchParams }: ChatDialogProps) {
   return (
     <Dialog open={isChatOpen}>
       <DialogContent className="w-full max-w-6xl h-[90vh]">
-        <DialogTitle>채팅</DialogTitle>
+        <DialogTitle></DialogTitle>
         <main className="w-full flex h-full gap-5 px-20">
           <div className="w-full flex gap-5 justify-center h-200">
             <ChatRoomList selectedRoomId={roomId} />
