+++ src/features/chat/ui/chatRoom/ChatRoomHeader.tsx
@@ -1,13 +1,13 @@
-import { Button } from '@/components/ui/button';
-import useChatRooms from '../hooks/useChatRooms';
+// import { Button } from '@/components/ui/button';
 import Image from 'next/image';
+// import useChatRooms from '../../hooks/useChatRooms';
 
 interface IChatRoomHeaderProps {
-  roomId: string;
+  // roomId: string;
   roomTitle: string;
 }
-export default function ChatRoomHeader({ roomId, roomTitle }: IChatRoomHeaderProps) {
-  const { deleteRoom } = useChatRooms();
+export default function ChatRoomHeader({ roomTitle }: IChatRoomHeaderProps) {
+  // const { deleteRoom } = useChatRooms();
 
   return (
     <div className="flex items-center justify-between p-4 border-b border-border_primary/20">
@@ -22,12 +22,12 @@ export default function ChatRoomHeader({ roomId, roomTitle }: IChatRoomHeaderPro
         <h3 className="font-medium text-white">{roomTitle}</h3>
       </div>
       <div className="flex items-center space-x-2">
-        <Button
+        {/* <Button
           onClick={() => deleteRoom(Number(roomId))}
           className="p-2 hover:bg-white/10 rounded-[10px] transition-colors"
         >
           삭제
-        </Button>
+        </Button> */}
       </div>
     </div>
   );
